Modularizando funciones de App.jsx
